Start moving PromQL queries to a central location
