+++ frontend/packages/console-shared/src/promql/OWNERS
@@ -1,0 +1,11 @@
+reviewers:
+  - kyoto
+  - lilic
+  - paulfantom
+  - s-urbaniak
+  - simonpasquier
+  - spadgett
+approvers:
+  - kyoto
+  - lilic
+  - s-urbaniak
