+++ frontend/packages/console-shared/src/promql/cluster-dashboard.ts
